requirements.txt
torch>=2.0.0
transformers>=4.30.0
pillow>=10.0.0
numpy>=1.24.0
clip-by-openai>=1.0
cairosvg>=2.7.0
python-lottie>=0.7.0

